+++ sdk/storage/storage-blob/src/Clients.ts
@@ -78,7 +78,8 @@ import {
   MatchConditions,
   ModificationConditions,
   ModifiedAccessConditions,
-  BlobQueryArrowField
+  BlobQueryArrowField,
+  BlobImmutabilityPolicy
 } from "./models";
 import {
   PageBlobGetPageRangesDiffResponse,
@@ -131,6 +132,11 @@ import { SasIPRange } from "./sas/SasIPRange";
 import { generateBlobSASQueryParameters } from "./sas/BlobSASSignatureValues";
 import { BlobSASPermissions } from "./sas/BlobSASPermissions";
 import { BlobLeaseClient } from "./BlobLeaseClient";
+import {
+  BlobDeleteImmutabilityPolicyResponse,
+  BlobSetImmutabilityPolicyResponse,
+  BlobSetLegalHoldResponse
+} from "./generatedModels";
 
 /**
  * Options to configure the {@link BlobClient.beginCopyFromURL} operation.
@@ -532,6 +538,18 @@ export interface BlobStartCopyFromURLOptions extends CommonOptions {
    */
   rehydratePriority?: RehydratePriority;
   /**
+   * Optional. Specifies immutability policy for a blob.
+   * Note that is parameter is only applicable to a blob within a container that
+   * has version level worm enabled.
+   */
+  immutabilityPolicy?: BlobImmutabilityPolicy;
+  /**
+   * Optional. Indicates if a legal hold should be placed on the blob.
+   * Note that is parameter is only applicable to a blob within a container that
+   * has version level worm enabled.
+   */
+  legalHold?: boolean;
+  /**
    * Blob tags.
    */
   tags?: Tags;
@@ -582,6 +600,18 @@ export interface BlobSyncCopyFromURLOptions extends CommonOptions {
    * Specify the md5 calculated for the range of bytes that must be read from the copy source.
    */
   sourceContentMD5?: Uint8Array;
+  /**
+   * Optional. Specifies immutability policy for a blob.
+   * Note that is parameter is only applicable to a blob within a container that
+   * has version level worm enabled.
+   */
+  immutabilityPolicy?: BlobImmutabilityPolicy;
+  /**
+   * Optional. Indicates if a legal hold should be placed on the blob.
+   * Note that is parameter is only applicable to a blob within a container that
+   * has version level worm enabled.
+   */
+  legalHold?: boolean;
   /**
    * Blob tags.
    */
@@ -757,6 +787,40 @@ export interface BlobGenerateSasUrlOptions extends CommonGenerateSasUrlOptions {
    * Optional only when identifier is provided. Specifies the list of permissions to be associated with the SAS.
    */
   permissions?: BlobSASPermissions;
+}
+
+/**
+ * Options for deleting immutability policy {@link BlobClient.deleteImmutabilityPolicy} operation.
+ */
+export interface BlobDeleteImmutabilityPolicyOptions extends CommonOptions {
+  /**
+   * An implementation of the `AbortSignalLike` interface to signal the request to cancel the operation.
+   * For example, use the &commat;azure/abort-controller to create an `AbortSignal`.
+   */
+  abortSignal?: AbortSignalLike;
+}
+
+/**
+ * Options for setting immutability policy {@link BlobClient.setImmutabilityPolicy} operation.
+ */
+export interface BlobSetImmutabilityPolicyOptions extends CommonOptions {
+  /**
+   * An implementation of the `AbortSignalLike` interface to signal the request to cancel the operation.
+   * For example, use the &commat;azure/abort-controller to create an `AbortSignal`.
+   */
+  abortSignal?: AbortSignalLike;
+  modifiedAccessCondition?: ModificationConditions;
+}
+
+/**
+ * Options for setting legal hold {@link BlobClient.setLegalHold} operation.
+ */
+export interface BlobSetLegalHoldOptions extends CommonOptions {
+  /**
+   * An implementation of the `AbortSignalLike` interface to signal the request to cancel the operation.
+   * For example, use the &commat;azure/abort-controller to create an `AbortSignal`.
+   */
+  abortSignal?: AbortSignalLike;
 }
 
 /**
@@ -1716,6 +1780,9 @@ export class BlobClient extends StorageClient {
         },
         sourceContentMD5: options.sourceContentMD5,
         blobTagsString: toBlobTagsString(options.tags),
+        immutabilityPolicyExpiry: options.immutabilityPolicy?.expiriesOn,
+        immutabilityPolicyMode: options.immutabilityPolicy?.policyMode,
+        legalHold: options.legalHold,
         ...convertTracingToRequestOptionsBase(updatedOptions)
       });
     } catch (e) {
@@ -2071,6 +2138,9 @@ export class BlobClient extends StorageClient {
           sourceIfUnmodifiedSince: options.sourceConditions.ifUnmodifiedSince,
           sourceIfTags: options.sourceConditions.tagConditions
         },
+        immutabilityPolicyExpiry: options.immutabilityPolicy?.expiriesOn,
+        immutabilityPolicyMode: options.immutabilityPolicy?.policyMode,
+        legalHold: options.legalHold,
         rehydratePriority: options.rehydratePriority,
         tier: toAccessTier(options.tier),
         blobTagsString: toBlobTagsString(options.tags),
@@ -2121,6 +2191,86 @@ export class BlobClient extends StorageClient {
       resolve(appendToURLQuery(this.url, sas));
     });
   }
+
+  /**
+   * Delete the immutablility policy on the blob.
+   *
+   * @param options - Optional options to delete immutability policy on the blob.
+   */
+  public async deleteImmutabilityPolicy(
+    options?: BlobDeleteImmutabilityPolicyOptions
+  ): Promise<BlobDeleteImmutabilityPolicyResponse> {
+    const { span, updatedOptions } = createSpan("BlobClient-deleteImmutabilityPolicy", options);
+    try {
+      return await this.blobContext.deleteImmutabilityPolicy({
+        abortSignal: options?.abortSignal,
+        ...convertTracingToRequestOptionsBase(updatedOptions)
+      });
+    } catch (e) {
+      span.setStatus({
+        code: SpanStatusCode.ERROR,
+        message: e.message
+      });
+      throw e;
+    } finally {
+      span.end();
+    }
+  }
+
+  /**
+   * Set immutablility policy on the blob.
+   *
+   * @param options - Optional options to set immutability policy on the blob.
+   */
+  public async setImmutabilityPolicy(
+    immutabilityPolicy: BlobImmutabilityPolicy,
+    options?: BlobSetImmutabilityPolicyOptions
+  ): Promise<BlobSetImmutabilityPolicyResponse> {
+    const { span, updatedOptions } = createSpan("BlobClient-setImmutabilityPolicy", options);
+    try {
+      return await this.blobContext.setImmutabilityPolicy({
+        abortSignal: options?.abortSignal,
+        immutabilityPolicyExpiry: immutabilityPolicy.expiriesOn,
+        immutabilityPolicyMode: immutabilityPolicy.policyMode,
+        modifiedAccessConditions: options?.modifiedAccessCondition,
+        ...convertTracingToRequestOptionsBase(updatedOptions)
+      });
+    } catch (e) {
+      span.setStatus({
+        code: SpanStatusCode.ERROR,
+        message: e.message
+      });
+      throw e;
+    } finally {
+      span.end();
+    }
+  }
+
+  /**
+   * Set legal hold on the blob.
+   *
+   * @param options - Optional options to set legal hold on the blob.
+   */
+  public async setLegalHold(
+    legalHoldEnabled: boolean,
+    options?: BlobSetLegalHoldOptions
+  ): Promise<BlobSetLegalHoldResponse> {
+    const { span, updatedOptions } = createSpan("BlobClient-setLegalHold", options);
+    try {
+      return await this.blobContext.setLegalHold(legalHoldEnabled, {
+        abortSignal: options?.abortSignal,
+        ...convertTracingToRequestOptionsBase(updatedOptions)
+      });
+    } catch (e) {
+      span.setStatus({
+        code: SpanStatusCode.ERROR,
+        message: e.message
+      });
+      throw e;
+    } finally {
+      span.end();
+    }
+  }
 }
 
 /**
@@ -2160,6 +2310,18 @@ export interface AppendBlobCreateOptions extends CommonOptions {
    */
   encryptionScope?: string;
   /**
+   * Optional. Specifies immutability policy for a blob.
+   * Note that is parameter is only applicable to a blob within a container that
+   * has version level worm enabled.
+   */
+  immutabilityPolicy?: BlobImmutabilityPolicy;
+  /**
+   * Optional. Indicates if a legal hold should be placed on the blob.
+   * Note that is parameter is only applicable to a blob within a container that
+   * has version level worm enabled.
+   */
+  legalHold?: boolean;
+  /**
    * Blob tags.
    */
   tags?: Tags;
@@ -2196,6 +2358,18 @@ export interface AppendBlobCreateIfNotExistsOptions extends CommonOptions {
    * Storage Services.
    */
   encryptionScope?: string;
+  /**
+   * Optional. Specifies immutability policy for a blob.
+   * Note that is parameter is only applicable to a blob within a container that
+   * has version level worm enabled.
+   */
+  immutabilityPolicy?: BlobImmutabilityPolicy;
+  /**
+   * Optional. Indicates if a legal hold should be placed on the blob.
+   * Note that is parameter is only applicable to a blob within a container that
+   * has version level worm enabled.
+   */
+  legalHold?: boolean;
 }
 
 /**
@@ -2516,6 +2690,9 @@ export class AppendBlobClient extends BlobClient {
         },
         cpkInfo: options.customerProvidedKey,
         encryptionScope: options.encryptionScope,
+        immutabilityPolicyExpiry: options.immutabilityPolicy?.expiriesOn,
+        immutabilityPolicyMode: options.immutabilityPolicy?.policyMode,
+        legalHold: options.legalHold,
         blobTagsString: toBlobTagsString(options.tags),
         ...convertTracingToRequestOptionsBase(updatedOptions)
       });
@@ -2771,6 +2948,18 @@ export interface BlockBlobUploadOptions extends CommonOptions {
    */
   tier?: BlockBlobTier | string;
   /**
+   * Optional. Specifies immutability policy for a blob.
+   * Note that is parameter is only applicable to a blob within a container that
+   * has version level worm enabled.
+   */
+  immutabilityPolicy?: BlobImmutabilityPolicy;
+  /**
+   * Optional. Indicates if a legal hold should be placed on the blob.
+   * Note that is parameter is only applicable to a blob within a container that
+   * has version level worm enabled.
+   */
+  legalHold?: boolean;
+  /**
    * Blob tags.
    */
   tags?: Tags;
@@ -2925,6 +3114,16 @@ export interface BlobQueryArrowConfiguration {
    * List of {@link BlobQueryArrowField} describing the schema of the data.
    */
   schema: BlobQueryArrowField[];
+}
+
+/**
+ * Options to query blob with Parquet format. Only valid for {@link BlockBlobQueryOptions.inputTextConfiguration}.
+ */
+export interface BlobQueryParquetConfiguration {
+  /**
+   * Kind.
+   */
+  kind: "parquet";
 }
 
 /**
@@ -3091,6 +3290,18 @@ export interface BlockBlobCommitBlockListOptions extends CommonOptions {
    * Storage Services.
    */
   encryptionScope?: string;
+  /**
+   * Optional. Specifies immutability policy for a blob.
+   * Note that is parameter is only applicable to a blob within a container that
+   * has version level worm enabled.
+   */
+  immutabilityPolicy?: BlobImmutabilityPolicy;
+  /**
+   * Optional. Indicates if a legal hold should be placed on the blob.
+   * Note that is parameter is only applicable to a blob within a container that
+   * has version level worm enabled.
+   */
+  legalHold?: boolean;
   /**
    * Access tier.
    * More Details - https://docs.microsoft.com/en-us/azure/storage/blobs/storage-blob-storage-tiers
@@ -3559,6 +3770,9 @@ export class BlockBlobClient extends BlobClient {
         },
         cpkInfo: options.customerProvidedKey,
         encryptionScope: options.encryptionScope,
+        immutabilityPolicyExpiry: options.immutabilityPolicy?.expiriesOn,
+        immutabilityPolicyMode: options.immutabilityPolicy?.policyMode,
+        legalHold: options.legalHold,
         tier: toAccessTier(options.tier),
         blobTagsString: toBlobTagsString(options.tags),
         ...convertTracingToRequestOptionsBase(updatedOptions)
@@ -3760,6 +3974,9 @@ export class BlockBlobClient extends BlobClient {
           },
           cpkInfo: options.customerProvidedKey,
           encryptionScope: options.encryptionScope,
+          immutabilityPolicyExpiry: options.immutabilityPolicy?.expiriesOn,
+          immutabilityPolicyMode: options.immutabilityPolicy?.policyMode,
+          legalHold: options.legalHold,
           tier: toAccessTier(options.tier),
           blobTagsString: toBlobTagsString(options.tags),
           ...convertTracingToRequestOptionsBase(updatedOptions)
@@ -4212,6 +4429,18 @@ export interface PageBlobCreateOptions extends CommonOptions {
    */
   encryptionScope?: string;
   /**
+   * Optional. Specifies immutability policy for a blob.
+   * Note that is parameter is only applicable to a blob within a container that
+   * has version level worm enabled.
+   */
+  immutabilityPolicy?: BlobImmutabilityPolicy;
+  /**
+   * Optional. Indicates if a legal hold should be placed on the blob.
+   * Note that is parameter is only applicable to a blob within a container that
+   * has version level worm enabled.
+   */
+  legalHold?: boolean;
+  /**
    * Access tier.
    * More Details - https://docs.microsoft.com/en-us/azure/storage/blobs/storage-blob-storage-tiers
    */
@@ -4255,6 +4484,18 @@ export interface PageBlobCreateIfNotExistsOptions extends CommonOptions {
    * Storage Services.
    */
   encryptionScope?: string;
+  /**
+   * Optional. Specifies immutability policy for a blob.
+   * Note that is parameter is only applicable to a blob within a container that
+   * has version level worm enabled.
+   */
+  immutabilityPolicy?: BlobImmutabilityPolicy;
+  /**
+   * Optional. Indicates if a legal hold should be placed on the blob.
+   * Note that is parameter is only applicable to a blob within a container that
+   * has version level worm enabled.
+   */
+  legalHold?: boolean;
   /**
    * Access tier.
    * More Details - https://docs.microsoft.com/en-us/azure/storage/blobs/storage-blob-storage-tiers
@@ -4667,6 +4908,9 @@ export class PageBlobClient extends BlobClient {
         },
         cpkInfo: options.customerProvidedKey,
         encryptionScope: options.encryptionScope,
+        immutabilityPolicyExpiry: options.immutabilityPolicy?.expiriesOn,
+        immutabilityPolicyMode: options.immutabilityPolicy?.policyMode,
+        legalHold: options.legalHold,
         tier: toAccessTier(options.tier),
         blobTagsString: toBlobTagsString(options.tags),
         ...convertTracingToRequestOptionsBase(updatedOptions)