+++ sdk/storage/storage-blob/src/ContainerClient.ts
@@ -521,6 +521,14 @@ export interface ContainerListBlobsOptions extends CommonOptions {
    * Specifies whether deleted blob with versions be returned in the response.
    */
   includeDeletedwithVersions?: boolean;
+  /**
+   * Specifies whether blob immutability policy be returned in the response.
+   */
+  includeImmutabilityPolicy?: boolean;
+  /**
+   * Specifies whether blob legal hold be returned in the response.
+   */
+  includeLegalHold?: boolean;
 }
 
 /**
@@ -1259,6 +1267,7 @@ export class ContainerClient extends StorageClient {
         ...options,
         ...convertTracingToRequestOptionsBase(updatedOptions)
       });
+
       const wrappedResponse: ContainerListBlobFlatSegmentResponse = {
         ...response,
         _response: response._response, // _response is made non-enumerable
@@ -1481,6 +1490,12 @@ export class ContainerClient extends StorageClient {
     if (options.includeDeletedwithVersions) {
       include.push("deletedwithversions");
     }
+    if (options.includeImmutabilityPolicy) {
+      include.push("immutabilitypolicy");
+    }
+    if (options.includeLegalHold) {
+      include.push("legalhold");
+    }
     if (options.prefix === "") {
       options.prefix = undefined;
     }
@@ -1691,6 +1706,12 @@ export class ContainerClient extends StorageClient {
     }
     if (options.includeDeletedwithVersions) {
       include.push("deletedwithversions");
+    }
+    if (options.includeImmutabilityPolicy) {
+      include.push("immutabilitypolicy");
+    }
+    if (options.includeLegalHold) {
+      include.push("legalhold");
     }
     if (options.prefix === "") {
       options.prefix = undefined;